+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ✅ test(maths-round): Round testcases and benchmarks added
 * 🎉 feat(maths-round): Round function added
 * ✅ test(maths): added two test cases for Floor and Ceil functions
 * 🐛 fix(maths-power): fixed returning float if output was float